pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "linkover"
version = "0.1.0"
description = "Pushover-based link sharing client for GNOME"
requires-python = ">=3.10"
dependencies = [
    "websocket-client>=1.6",
    "requests>=2.28",
    "rumps>=0.4; sys_platform == 'darwin'",
    "Pillow>=9.0; sys_platform == 'darwin'",
]

[project.scripts]
linkover = "linkover.__main__:main"

[tool.hatch.build.targets.wheel]
include = ["linkover/**"]
